Move functions to match directory with /path/to/endpoint
