+++ services/src/backgrounds/backgrounds.ts
@@ -1,12 +1,12 @@
 import { unsplashImagesCollections, unsplashImagesSearch } from './unsplash/images/user.ts'
-import { storeDaylightVideos } from './bonjourr/videos/store.ts'
+import { storeDaylightImages } from './bonjourr/images/store/store.ts'
+import { storeDaylightVideos } from './bonjourr/videos/store/store.ts'
 import { pixabayVideosSearch } from './pixabay/videos/search.ts'
 import { pixabayImagesSearch } from './pixabay/images/search.ts'
-import { storeDaylightImages } from './bonjourr/images/store.ts'
 import { metMuseumPaintings } from './metmuseum/images/paintings.ts'
 import { getDaylightImages } from './bonjourr/images/get.ts'
 import { getDaylightVideos } from './bonjourr/videos/get.ts'
-import { getAllStoredMedia } from './bonjourr/all.ts'
+import { getAllStoredMedia } from './bonjourr/all/all.ts'
 import { initUnsplashAuth } from './unsplash/shared.ts'
 import { backgroundsProxy } from './proxy.ts'
 import { metMuseumSearch } from './metmuseum/images/search.ts'
