+++ services/src/backgrounds/bonjourr/videos/store/store-pexels.ts
@@ -1,9 +1,9 @@
-import { pexelsToGeneric } from '../../pexels/convert.ts'
+import { pexelsToGeneric } from '../../../pexels/convert.ts'
 
-import type { PexelsCollection } from '../../pexels/types.ts'
-import type { CollectionList } from '../shared.ts'
-import type { Video } from '../../types.ts'
-import type { Env } from '../../../index.ts'
+import type { PexelsCollection } from '../../../pexels/types.ts'
+import type { CollectionList } from '../../shared.ts'
+import type { Video } from '../../../types.ts'
+import type { Env } from '../../../../index.ts'
 
 export const PEXELS_COLLECTIONS = {
 	'bonjourr-videos-daylight-night': 'ohpcuko',
