+++ services/src/backgrounds/metmuseum/images/search.ts
@@ -1,5 +1,5 @@
 import { fetchSingleObject, metObjectToBonjourr } from '../shared.ts'
-import type { MetObject, MetObjectsIds } from '../shared.ts'
+import type { MetObject, MetObjectsIds } from '../types.ts'
 
 export async function metMuseumSearch(url: URL, headers: Headers): Promise<Response> {
 	const amount = Math.min(Number.parseInt(url.searchParams.get('amount') ?? '10'), 40)
